feat: some refactor
