+++ internal/jwt/jwt.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"go-template/models"
-	resultwrapper "go-template/pkg/utl/result_wrapper"
+	resultwrapper "go-template/pkg/utl/resultwrapper"
 
 	jwt "github.com/dgrijalva/jwt-go"
 	"github.com/volatiletech/sqlboiler/v4/boil"
