+++ internal/middleware/auth/auth.go
@@ -6,7 +6,7 @@ import (
 
 	"go-template/daos"
 	"go-template/models"
-	resultwrapper "go-template/pkg/utl/result_wrapper"
+	resultwrapper "go-template/pkg/utl/resultwrapper"
 
 	graphql2 "github.com/99designs/gqlgen/graphql"
 	jwt "github.com/dgrijalva/jwt-go"
